+++ front/src/app/pages/Alunos/turmas/templates/ListagemTurma.tsx
@@ -1,9 +1,23 @@
-    import { Table, Badge, Button, ButtonGroup } from 'react-bootstrap';
-    import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-    import { faEdit, faTrash, faPowerOff } from '@fortawesome/free-solid-svg-icons';
+import React, { useState } from "react";
+import { Table, Badge, Button, ButtonGroup } from 'react-bootstrap';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import {faCheck, faEdit, faPowerOff} from '@fortawesome/free-solid-svg-icons';
+import Modal from "react-bootstrap/esm/Modal";
+import EditarTurma from "./EditarTurma";
+import InativarTurma from "./InativarTurma";
 
-    const ListagemTurma: React.FC = () => {
-        return (
+const ListagemTurma: React.FC = () => {
+
+    const [showEditar, setShowEditar] = useState(false);
+    const handleCloseEditar = () => setShowEditar(false);
+    const handleShowEditar = () => setShowEditar(true);
+
+    const [showInativar, setShowInativar] = useState(false);
+    const handleCloseInativar = () => setShowInativar(false);
+    const handleShowInativar = () => setShowInativar(true);
+
+    return (
+        <>
             <Table striped className="tabela">
                 <thead>
                 <tr>
@@ -24,10 +38,10 @@
                     </td>
                     <td>
                         <ButtonGroup aria-label="Ações" className="tabela-acoes">
-                            <Button variant="btn-outline-secondary" className="color-green">
+                            <Button onClick={handleShowEditar} variant="btn-outline-secondary" className="color-green">
                                 <FontAwesomeIcon icon={faEdit} />
                             </Button>
-                            <Button variant="btn-outline-secondary" className="color-red">
+                            <Button onClick={handleShowInativar} variant="btn-outline-secondary" className="color-red">
                                 <FontAwesomeIcon icon={faPowerOff} />
                             </Button>
                         </ButtonGroup>
@@ -42,10 +56,10 @@
                     </td>
                     <td>
                         <ButtonGroup aria-label="Ações" className="tabela-acoes">
-                            <Button variant="btn-outline-secondary" className="color-green">
+                            <Button onClick={handleShowEditar} variant="btn-outline-secondary" className="color-green">
                                 <FontAwesomeIcon icon={faEdit} />
                             </Button>
-                            <Button variant="btn-outline-secondary" className="color-red">
+                            <Button onClick={handleShowInativar} variant="btn-outline-secondary" className="color-red">
                                 <FontAwesomeIcon icon={faPowerOff} />
                             </Button>
                         </ButtonGroup>
@@ -60,10 +74,10 @@
                     </td>
                     <td>
                         <ButtonGroup aria-label="Ações" className="tabela-acoes">
-                            <Button variant="btn-outline-secondary" className="color-green">
+                            <Button onClick={handleShowEditar} variant="btn-outline-secondary" className="color-green">
                                 <FontAwesomeIcon icon={faEdit} />
                             </Button>
-                            <Button variant="btn-outline-secondary" className="color-red">
+                            <Button onClick={handleShowInativar} variant="btn-outline-secondary" className="color-red">
                                 <FontAwesomeIcon icon={faPowerOff} />
                             </Button>
                         </ButtonGroup>
@@ -71,7 +85,58 @@
                 </tr>
                 </tbody>
             </Table>
-        );
-    };
 
-    export default ListagemTurma;
+            <Modal
+                show={showEditar}
+                onHide={handleCloseEditar}
+                size="lg"
+                backdrop="static"
+                centered
+                keyboard={false}>
+
+                <Modal.Header closeButton>
+                    <Modal.Title>Ediar Turma</Modal.Title>
+                </Modal.Header>
+
+                <Modal.Body>
+                    <EditarTurma/>
+                </Modal.Body>
+
+                <Modal.Footer>
+                    <Button variant="secondary" onClick={handleCloseEditar}>Desistir</Button>
+                    <Button variant="success">
+                        <FontAwesomeIcon icon={faCheck} /> Salvar
+                    </Button>
+
+                </Modal.Footer>
+            </Modal>
+
+            <Modal
+                show={showInativar}
+                onHide={handleCloseInativar}
+                size="sm"
+                backdrop="static"
+                centered
+                keyboard={false}>
+
+                <Modal.Header closeButton>
+                    <Modal.Title>Confirmação</Modal.Title>
+                </Modal.Header>
+
+                <Modal.Body>
+                    <InativarTurma/>
+                </Modal.Body>
+
+                <Modal.Footer>
+                    <Button variant="secondary" onClick={handleCloseInativar}>Desistir</Button>
+                    <Button variant="danger">
+                        <FontAwesomeIcon icon={faCheck} /> Inativar
+                    </Button>
+
+                </Modal.Footer>
+            </Modal>
+        </>
+    );
+};
+
+export default ListagemTurma;
